requirements.txt
torch==1.13
datasets==2.15.0
transformers==4.25.1
tqdm
pandas
scikit-learn

# faiss install (if you want to)
faiss-gpu

#
omegaconf==2.3.0
wandb==0.18.3
hydra-core==1.3.0